Rename bspnode.h to bsptree.h
