+++ genlib/bsptree.h
